Funcionalidad, Shortcuts, Certificados

+++ resources/views/personal/personal.blade.php
@@ -13,15 +13,22 @@
                 @include('alerts.request')
                 @include('alerts.errors')
             </div>
-
+        @if(verificarPeriodo()=='activo')
             <div class="col-xs-12">
                 <button class="btn btn-primary" title="Registrar personal" onclick="window.location.href = '{{ URL::to('/nuevo_personal') }}'";>
                     <span class="glyphicon glyphicon-plus" aria-hidden="true"></span>Nuevo
                 </button>
             </div>
-
+        @endif
             <div class="col-xs-12" style="padding-top: 20px">
-                <div class="box">
+                   <!-- <div class="widget-box">
+          <div class="widget-title"> <span class="icon"><i class="icon-th"></i></span>
+            <h5>Listado de Personal</h5>
+          </div>
+          <div class="table-responsive">
+          <div class="widget-content nopadding">
+            <table class="table table-bordered data-table">   -->
+            <div class="box">
                     <div class="box-header">
                         <h3 class="box-title">Personal</h3>
                     </div>
@@ -47,22 +54,26 @@
                                             <td>{{$per->cedula}}</td>
                                             <td>{{$per->cargo->nombre}}</td>
                                             <td>{{$per->cargo->area->nombre}}</td>
+                                    @if(verificarPeriodo()=='activo')
                                             <td>
                                                 <!-- <a href="{{ route('personal.show', $per->id) }}" class="btn btn-default btn-flat"><i class="fa fa-eye"></i></a> -->
                                                  <a href="{{ route('personal.edit', $per->id) }}" class="btn btn-primary btn-flat"><i class="fa fa-refresh"></i></a>
                                                 <a href="personal/{{ $per->id }}/destroy" class="btn btn-danger btn-flat"><i class="fa fa-trash"></i></a>
                                             </td>
-
+                                    @else
+                                            <td></td>
+                                    @endif
                                         </tr>
+                                        
                                     @endforeach
                                     </tbody>
                                 </table>
                             </div>
                         </div>
+                        </div>
                     </div>
                 </div>
             </div>
         </div>
-    </div>
 
 @endsection
